Removed duplicate properties (#12080)
* Removed o_ prefix from o_properties property

* Removed internal o_properties and moved setProperty to AbstractElement

* Removed duplicate code from Asset and Document classes

* Removed duplicate code

* Resolved object type by service
