+++ models/Document.php
@@ -148,13 +148,6 @@ class Document extends Element\AbstractElement
     /**
      * @internal
      *
-     * @var array|null
-     */
-    protected $properties = null;
-
-    /**
-     * @internal
-     *
      * @var array
      */
     protected $children = [];
@@ -1246,59 +1239,6 @@ class Document extends Element\AbstractElement
     public function setPublished($published)
     {
         $this->published = (bool) $published;
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getProperties()
-    {
-        if ($this->properties === null) {
-            // try to get from cache
-            $cacheKey = 'document_properties_' . $this->getId();
-            $properties = \Pimcore\Cache::load($cacheKey);
-            if (!is_array($properties)) {
-                $properties = $this->getDao()->getProperties();
-                $elementCacheTag = $this->getCacheTag();
-                $cacheTags = ['document_properties' => 'document_properties', $elementCacheTag => $elementCacheTag];
-                \Pimcore\Cache::save($properties, $cacheKey, $cacheTags);
-            }
-
-            $this->setProperties($properties);
-        }
-
-        return $this->properties;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setProperties(?array $properties)
-    {
-        $this->properties = $properties;
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setProperty($name, $type, $data, $inherited = false, $inheritable = false)
-    {
-        $this->getProperties();
-
-        $property = new Property();
-        $property->setType($type);
-        $property->setCid($this->getId());
-        $property->setName($name);
-        $property->setCtype('document');
-        $property->setData($data);
-        $property->setInherited($inherited);
-        $property->setInheritable($inheritable);
-
-        $this->properties[$name] = $property;
 
         return $this;
     }
